requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2
django-allauth==0.50.0
django-cors-headers==3.13.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
idna==3.4
joblib==1.2.0
numpy==1.23.4
oauthlib==3.2.2
pandas==1.5.1
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
urllib3==1.26.12
